# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/survey/category/{categoryId}": {
      "get": {
        "description": "获取指定分类下进行中的问卷",
        "operationId": "PublicSurveyController_findByCategory",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回问卷列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicSurveyListItemVO"
                  }
                }
              }
            }
          }
        },
        "summary": "按分类获取问卷",
        "tags": [
          "问卷 - 用户端"
        ]
      }
    },
    "/survey/detail/{uid}": {
      "get": {
        "description": "根据 UID 获取问卷详情，用于答题",
        "operationId": "PublicSurveyController_findOne",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "问卷 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回问卷详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSurveyDetailVO"
                }
              }
            }
          }
        },
        "summary": "获取问卷详情",
        "tags": [
          "问卷 - 用户端"
        ]
      }
    },
...
    "/admin/survey/list": {
      "get": {
        "operationId": "SurveyController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，默认 20，最大 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "active",
                "closed"
              ]
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "description": "是否已归档",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "createdBy",
            "required": false,
            "in": "query",
            "description": "创建人 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字搜索（标题）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "required": false,
            "in": "query",
            "description": "开始时间（创建时间范围）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": false,
            "in": "query",
            "description": "结束时间（创建时间范围）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "是否包含已删除",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "id",
                "createdAt",
                "updatedAt",
                "sortOrder",
                "submitCount",
                "viewCount"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分页查询问卷列表",
        "tags": [
          "问卷管理"
        ]
      }
    },
...
    "/survey-response/submit": {
      "post": {
        "description": "提交问卷答案，需要登录。如果问卷设置了登录必填，则必须登录后才能提交。",
        "operationId": "SurveyResponseController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResultVO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "提交问卷响应",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/my-list": {
      "get": {
        "description": "查询当前登录用户的所有问卷响应",
        "operationId": "SurveyResponseController_findMyResponses",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，默认 20，最大 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "surveyUid",
            "required": false,
            "in": "query",
            "description": "问卷 UID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "响应状态",
            "schema": {
              "type": "string",
              "enum": [
                "submitted",
                "reviewing",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "isEffective",
            "required": false,
            "in": "query",
            "description": "是否有效",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回响应列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseListVO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询我的响应列表",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/detail/{uid}": {
      "get": {
        "description": "根据响应 UID 查询详情，只能查看自己的响应",
        "operationId": "SurveyResponseController_findByUid",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "响应 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回响应详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDetailVO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询响应详情",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/status/{surveyUid}": {
      "get": {
        "description": "查询当前用户在指定问卷的提交状态（是否已提交、提交次数等）",
        "operationId": "SurveyResponseController_getUserSurveyStatus",
        "parameters": [
          {
            "name": "surveyUid",
            "required": true,
            "in": "path",
            "description": "问卷 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回提交状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSurveyStatusVO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询问卷提交状态",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/update/{uid}": {
      "post": {
        "description": "更新问卷响应，只能更新自己的响应，且仅限 submitted 状态",
        "operationId": "SurveyResponseController_update",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "响应 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveyResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponseDetailVO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新响应",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/delete/{uid}": {
      "post": {
        "description": "删除问卷响应（软删除），只能删除自己的响应，且仅限 submitted 状态",
        "operationId": "SurveyResponseController_remove",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "响应 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除响应",
        "tags": [
          "问卷响应 - 用户端"
        ]
      }
    },
    "/survey-response/anonymous/submit": {
      "post": {
        "description": "匿名提交问卷答案。如果问卷设置了登录必填，则会返回 403 错误。",
        "operationId": "AnonymousSurveyResponseController_submitAnonymous",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "提交成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitResultVO"
                }
              }
            }
          }
        },
        "summary": "匿名提交问卷",
        "tags": [
          "问卷响应 - 匿名提交"
        ]
      }
    }
  },
  "info": {
    "title": "jmni-server",
    "description": "JMNI Server API 文档 - 供外部团队参考",
    "version": "0.0.1",
    "contact": {
      "name": "JMNI Team",
      "url": "https://jmni.cn",
      "email": "support@jmni.cn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://api.jmni.cn",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:2233",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Access Token"
      }
    },
    "schemas": {
...
      "PublicSurveyListItemVO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "问卷 UID"
          },
          "title": {
            "type": "object",
            "description": "问卷标题（多语言）"
          },
          "description": {
            "type": "object",
            "description": "问卷描述（多语言）"
          },
          "themeColor": {
            "type": "string",
            "description": "主题主色"
          },
          "loginRequired": {
            "type": "boolean",
            "description": "是否需要登录"
          },
          "startTime": {
            "type": "string",
            "description": "开始时间"
          },
          "endTime": {
            "type": "string",
            "description": "截止时间"
          },
          "categoryId": {
            "type": "number",
            "description": "分类 ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          }
        },
        "required": [
          "uid",
          "title",
          "loginRequired",
          "categoryName"
        ]
      },
      "PublicSurveyDetailVO": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "问卷 UID"
          },
          "title": {
            "type": "object",
            "description": "问卷标题（多语言）"
          },
          "description": {
            "type": "object",
            "description": "问卷描述（多语言）"
          },
          "themeColor": {
            "type": "string",
            "description": "主题主色"
          },
          "loginRequired": {
            "type": "boolean",
            "description": "是否需要登录"
          },
          "startTime": {
            "type": "string",
            "description": "开始时间"
          },
          "endTime": {
            "type": "string",
            "description": "截止时间"
          },
          "categoryId": {
            "type": "number",
            "description": "分类 ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "topics": {
            "type": "object",
            "description": "问卷结构（题目列表配置 JSON schema）"
          },
          "endMessage": {
            "type": "object",
            "description": "结束语（多语言）"
          },
          "showQuestionIndex": {
            "type": "boolean",
            "description": "是否显示题目编号"
          },
          "languagesList": {
            "description": "启用的多语言列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requireGameBinding": {
            "type": "boolean",
            "description": "是否要求绑定游戏账号"
          },
          "maxSubmitTimesPerUser": {
            "type": "number",
            "description": "每用户最大提交次数"
          }
        },
        "required": [
          "uid",
          "title",
          "loginRequired",
          "categoryName",
          "showQuestionIndex",
          "requireGameBinding",
          "maxSubmitTimesPerUser"
        ]
      },
...
      "CreateSurveyResponseDto": {
        "type": "object",
        "properties": {
          "surveyUid": {
            "type": "string",
            "description": "问卷 UID",
            "example": "srv_123456789012"
          },
          "answers": {
            "type": "object",
            "description": "问卷答案（JSON 对象，key 为题目 ID，value 为答案）",
            "example": {
              "q1": "A",
              "q2": [
                "B",
                "C"
              ],
              "q3": "这是我的回答"
            }
          },
          "durationSeconds": {
            "type": "number",
            "description": "填写时长（秒）",
            "example": 120,
            "minimum": 0
          },
          "locale": {
            "type": "string",
            "description": "提交者所选语言（如 zhCN, enUS）",
            "example": "zhCN",
            "maxLength": 10
          },
          "surveyLanguage": {
            "type": "string",
            "description": "提交时使用的问卷语言",
            "example": "zhCN",
            "maxLength": 10
          },
          "referrer": {
            "type": "string",
            "description": "来源 Referrer",
            "maxLength": 512
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称（自填）",
            "example": "Player123",
            "maxLength": 50
          },
          "guid": {
            "type": "string",
            "description": "用户 KID/GUID（游戏账号标识）",
            "example": "guid_abc123",
            "maxLength": 64
          },
          "gamelink": {
            "type": "object",
            "description": "用户游戏链接信息",
            "example": {
              "platform": "steam",
              "region": "AS",
              "uid": "76561198012345678"
            }
          },
          "email": {
            "type": "string",
            "description": "用户邮箱（自填）",
            "example": "user@example.com"
          },
          "timeZone": {
            "type": "string",
            "description": "提交者时区",
            "example": "Asia/Shanghai",
            "maxLength": 40
          }
        },
        "required": [
          "surveyUid",
          "answers"
        ]
      },
      "SubmitResultVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "responseUid": {
            "type": "string",
            "description": "响应 UID",
            "example": "rsp_123456789012"
          },
          "message": {
            "type": "string",
            "description": "消息"
          }
        },
        "required": [
          "success",
          "responseUid"
        ]
      },
      "SurveyResponseListItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "响应 ID",
            "example": 1
          },
          "uid": {
            "type": "string",
            "description": "响应 UID",
            "example": "rsp_123456789012"
          },
          "surveyUid": {
            "type": "string",
            "description": "问卷 UID",
            "example": "srv_123456789012"
          },
          "status": {
            "type": "string",
            "description": "响应状态",
            "enum": [
              "submitted",
              "reviewing",
              "approved",
              "rejected"
            ]
          },
          "isEffective": {
            "type": "boolean",
            "description": "是否有效",
            "example": true
          },
          "durationSeconds": {
            "type": "number",
            "description": "填写时长（秒）"
          },
          "locale": {
            "type": "string",
            "description": "提交语言"
          },
          "createdAt": {
            "type": "string",
            "description": "提交时间（ISO 8601）"
          }
        },
        "required": [
          "id",
          "uid",
          "surveyUid",
          "status",
          "isEffective",
          "createdAt"
        ]
      },
      "SurveyResponseListVO": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "总记录数",
            "example": 100
          },
          "items": {
            "description": "响应列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SurveyResponseListItemVO"
            }
          },
          "page": {
            "type": "number",
            "description": "当前页码",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "总页数",
            "example": 5
          }
        },
        "required": [
          "total",
          "items",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
      "SurveyResponseDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "响应 ID",
            "example": 1
          },
          "uid": {
            "type": "string",
            "description": "响应 UID",
            "example": "rsp_123456789012"
          },
          "surveyUid": {
            "type": "string",
            "description": "问卷 UID",
            "example": "srv_123456789012"
          },
          "status": {
            "type": "string",
            "description": "响应状态",
            "enum": [
              "submitted",
              "reviewing",
              "approved",
              "rejected"
            ]
          },
          "isEffective": {
            "type": "boolean",
            "description": "是否有效",
            "example": true
          },
          "durationSeconds": {
            "type": "number",
            "description": "填写时长（秒）"
          },
          "locale": {
            "type": "string",
            "description": "提交语言"
          },
          "createdAt": {
            "type": "string",
            "description": "提交时间（ISO 8601）"
          },
          "answers": {
            "type": "object",
            "description": "问卷答案"
          },
          "surveyLanguage": {
            "type": "string",
            "description": "问卷语言"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间（ISO 8601）"
          }
        },
        "required": [
          "id",
          "uid",
          "surveyUid",
          "status",
          "isEffective",
          "createdAt",
          "answers",
          "updatedAt"
        ]
      },
      "UserSurveyStatusVO": {
        "type": "object",
        "properties": {
          "surveyUid": {
            "type": "string",
            "description": "问卷 UID"
          },
          "hasSubmitted": {
            "type": "boolean",
            "description": "是否已提交",
            "example": false
          },
          "submitCount": {
            "type": "number",
            "description": "已提交次数"
          },
          "maxSubmitTimes": {
            "type": "number",
            "description": "最大可提交次数（0 表示不限）"
          },
          "canSubmit": {
            "type": "boolean",
            "description": "是否可以继续提交",
            "example": true
          },
          "lastResponseUid": {
            "type": "string",
            "description": "最后一次提交的响应 UID"
          },
          "lastSubmitTime": {
            "type": "string",
            "description": "最后提交时间"
          }
        },
        "required": [
          "surveyUid",
          "hasSubmitted",
          "canSubmit"
        ]
      },
      "UpdateSurveyResponseDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "object",
            "description": "问卷答案（JSON 对象）"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称",
            "maxLength": 50
          },
          "guid": {
            "type": "string",
            "description": "用户 KID/GUID",
            "maxLength": 64
          },
          "gamelink": {
            "type": "object",
            "description": "游戏链接信息"
          },
          "email": {
            "type": "string",
            "description": "用户邮箱"
          }
        }
      }
    }
  }
}